refactor(favorite-locations-pane): Rename file for component.
