+++ lib/components/user/favorite-locations-pane.js
