+++ lib/components/user/user-account-screen.js
@@ -10,7 +10,7 @@ import AppMenu from '../app/app-menu'
 import AccountCommands from './account-commands'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import AwaitingScreen from './awaiting-screen'
-import FavoriteLocationsPane from './favorite-places-pane'
+import FavoriteLocationsPane from './favorite-locations-pane'
 import NotificationPrefsPane from './notification-prefs-pane'
 import PhoneVerificationPane from './phone-verification-pane'
 import TermsOfUsePane from './terms-of-use-pane'
